scripts/spread_amx/requirements.txt
kafka-python==2.0.2
pyyaml==5.4.1
pysftp==0.2.9
